Support e2e on single node deployments
Currently, in e2e, we schedule pods sometimes on gateway nodes and sometimes
on non-gateway nodes to cover a wide range of use-cases. However, when the
same tests are executed on a single node deployments it returns an unfriendly
error message which would be hard to understand. This PR modifies the code
to run only the applicable test-scenarios.

Depends on: https://github.com/submariner-io/shipyard/pull/654
Signed-Off-by: Sridhar Gaddam <sgaddam@redhat.com>